requirements.txt
flask
crewai
twilio
groq
python-dotenv
unidecode